+++ bootique-agrest-junit5/src/main/java/io/bootique/agrest/junit5/AgResponseMatcher.java
@@ -17,7 +17,9 @@ import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * @since 2.0
+ * @deprecated The users are encouraged to switch to Jakarta-based flavor
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class AgResponseMatcher {
 
     private static final Pattern NUMERIC_ID_MATCHER = Pattern.compile("\"id\":([\\d]+)");