+++ bootique-agrest5-swagger/src/main/java/io/bootique/agrest/v5/swagger/AgrestSwaggerModuleExtender.java
@@ -27,7 +27,9 @@ import io.bootique.di.Binder;
 
 /**
  * @since 3.0
+ * @deprecated The users are encouraged to switch to Jakarta-based flavor
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class AgrestSwaggerModuleExtender extends ModuleExtender<AgrestSwaggerModuleExtender> {
 
     private AgrestModuleExtender agrestExtender;