+++ bootique-agrest5/src/main/java/io/bootique/agrest/v5/AgrestModule.java
@@ -37,7 +37,9 @@ import java.util.Set;
 
 /**
  * @since 3.0
+ * @deprecated The users are encouraged to switch to Jakarta-based flavor
  */
+@Deprecated(since = "3.0", forRemoval = true)
 public class AgrestModule extends ConfigModule {
 
     /**